requirements.txt
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
torch==2.1.0
torchvision==0.16.0
peft==0.7.1
bitsandbytes==0.41.3
gradio==4.16.0
gradio-client==0.8.1
pillow==10.1.0
numpy==1.26.4
safetensors==0.4.1
huggingface-hub==0.20.0
scipy